Add methods to retrieve the raw value of an indexed field
This commit adds code to get the index field value based on
FieldDefiniton. This can be used when index implementations
do not generally reconstruct data from the index but use a loading cache
instead (like GroupIndex) and the indexed value is of interest (e.g.
when detecting staleness of indexed documents).

We provide an abstraction over the indexed document called FieldsBundle
to make the approach work on different index backends. The index
backends can use the field definitions of the current search version to
help identify the data type.

This commit provides an implementation for both the ElasticSearch and
the Lucene backend for {Account,Group,Project,Change}Index.

Change-Id: I461067a3610b4e31e90519522004f047c54a838a