AbstractElasticIndex: Factor out more variants of {post,perform}Request
Factor out more variants of these methods that eventually end up in the
one that actually makes the call to the RestClient instance. This will
allow to make any necessary adjustments to the URI or request parameters
in a single place.

At the same time, reorder the arguments to make the order consistent
across methods.

Change-Id: I3ced32e8df6cbd5a5f05f4d3241f84a7e7bebb10