De-couple online reindex activation from index module
Gerrit plugins can contribute to submit rules evaluation, that are a
part of secondary index.  Therefore it is crucial that plugins are
loaded prior to triggering of online reindexing.

This change de-couples online reindex activation from index module and
thus postpones the triggering of online reindexing until after loading
of the plugins:

1. Load secondary index
2. Load all plugins
3. Trigger online reindex if needed

Bug: Issue 10082
Change-Id: I1706d42cb50eb342c2b9eca076e2def5bd1426c0