Consolidate base64 encoding/decoding to use Guava's BaseEncoding
Replace usage of Java's, JGit's and Apache Commons' Base64 classes with
BaseEncoding from Guava.

No Gerrit code directly depends on Apache Commons Codec any more, so we
can remove all the explicit dependencies from the build rules. However,
it is still a transient dependency of the httpclient library, and is
used by some plugins, so we can't completely remove it from the build.

Change-Id: I9a934bda3e4519f359759b20efb298d544f1391b