ElasticChangeIndex: Fix decoding of attention set field
The attention set field must be decoded from base64, in the same way
that the submit records field is.

Factor out the decoding of submit records to a separate method, and
reuse it to decode the attention set.

Re-enable the tests for V5, V6 and V7 change queries, which now pass.

Bug: Issue 12524
Change-Id: Ie5268df4b718733f032ddfeac86c1a0432948fe2